♻️ refactor: clean up type, create db type folder
